Use latest charabia
I'm trying to build odict with Nix. To build one of the
transitive dependencies, lindera, which charabia depends on,
I needed to prebuild its data artifacts so that it doesn't
try to download files from the internet when building.

The prebuilt cache only has artifacts for v0.32.2,
and charabia doesn't have a tagged release that depends
on that version of lindera.